Allow stripping a realm from user names before group lookups.
When AD/Kerberos authentication is used with squid, squid may pass the
user name with the kerberos realm appended. Depending on other setup
(Samba,LDAP ...), that realm may not appear in user names found when
searching for the users group membership.

The realm to strip can be specified using the '-r <realm>' command line
argument.