SG as suid ACL helper, optional reverse lookups.
When SG is started with the '-z' option, it acts as an ACL helper for squid
instead of acting as an redirector. This feature needs more work though, it is
not yet possible to pass more information (error messages, tags) back to squid.

The URL parser has been rewritten to allow it to be used to parse the
redirector format as well as the new format for an authorization helper.

It is now possible to do DNS reverse lookups for IPv4 and IPv6 addresses found
in the URL. This feature can be enabled by specifying the "reverselookup"
keyword in the configuration.