Support for tag=, log= and mesage= in authz mode.
The new 'tag'  keyword followed by a string can be used to assign a tag the
to the rule. The tag is then passed in the authz result back to squid.

"log=" and "message=" keywords are added the authz response, to give squid a
hint on the reason for the access decission.

HTParse.c again contains the HTEscape() function from the original libwww code,
albeit a bit modified to fit into SG. Also moved the few necessary pieces from
wwwsys.h into HTParse.c and removed wwwsys.h entirely.

The emergency mode is now handled in the main loop.