Implement error redirect.
It is now possible to specify an error redirect URL,  that is used when SG is
in emergency mode or when the request from squid cannot be parsed.

The redirect can be specified in the configuration or as command-line parameter.
Because errors parsing the configuration also cause SG to go into emergency mode,
the command-line method should be preferred.

The message will be appended to the URL as query parameter.

When no pass-through is set and no error redirect is specified, "Error" is
printed on stdout, what should make squid complain instead.