Unified ident handling in parse*Line, realm stripping.
It is now possible to strip the authentication realm (everything after
the last '@') from the users identity: When the "strip-realm" setting is set,
any realm gets stripped, when the "realm" setting is specified, the realm
is stripped when it matches the specified string.

This is useful in cases where group/netgroup/LDAP authorization is used in
conjunction with Kerberos or Samba integration.

One could also keep the stripped realm in the request structure and write a
source match to check for specific realms. Left that as an exerciser for
someone else :-)

Identity handling in parseLine() and parseAuthzLine() now uses a common
function instead of duplicated code.

Note that parseLine() now always un-escapes the identity field too, the old
code did that only when compiled with LDAP support - for whatever reason.