Shortcut for multiple destination matches.
Beause of the new ACL semantics, it is rather likely that the same request is
matched against against the same destination list multiple times, which can
be rather expensive with live checks. But since the request URL won't change,
the match result won't either, so we can simply shortcut the match.

The bypass is done by checking the (new) request serial number against the last
request seen by the destination list.